\name{coast}
\docType{data}
\alias{coast}
\title{Shapefile of the Irish and UK coastlines}
\description{
Shapefile list object (see: \code{\link[shapefiles]{shapefiles}}) of the Irish and UK coastline.
}
\usage{coast}
\format{shapefile list object}
\source{GSHHS A Global Self-consistent, Hierarchical, High-resolution Shoreline Database.}
\references{
\url{http://www.ngdc.noaa.gov/mgg/shorelines/gshhs.html}
}