\name{effort}
\docType{data}
\alias{effort}
\title{Spatially disaggregated fishing effort and landings data}
\description{
Monthly fishing effort and haddock landings landings at a spatial resolution of 1 degree longitude and 0.5 degrees latitude. The data are from Irish otter trawlers in 2009.
}
\usage{effort}
\format{data frame}
\source{EU Logbooks database}
