\name{landings}
\docType{data}
\alias{landings}
\title{Spatially disaggregated landings data}
\description{
Landings data of some gadoid species (cod, haddock, hake, saithe and whiting) at a spatial resolution of 1 degree longitude and 0.5 degrees latitude. The data are from Irish otter trawlers in 2009.
}
\usage{landings}
\format{data frame}
\source{Logbooks database}
